+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 - Permanantly moved original gregium to gregium.old, creating new readme
 - Ported content of gregium.old to new gregium
 - Recursivley found packages and packagedata in setup
+
+## 0.1.10
+
+- Resolved Dependecies
+
+## 0.1.11
+
+- Added easing module